requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
flask==3.0.0
pyyaml==6.0.1
schedule==1.2.0
python-dotenv==1.0.0
jinja2==3.1.3
markupsafe==2.1.5
